Add files via upload
ShoppingCart.java
•	Added the ShoppingCart class
•	Added private cart attribute (ArrayList<int[]>).
o	An ArrayList of integer arrays of length 2. Index 0 in each array is the product ID, and index 1 is the quantity in the cart.
•	Added public getCart() method that returns the cart attribute.
•	Added public method addProduct(int id, int quantity) that adds a product to the cart with the given quantity.
o	Throws an InvalidParameterException if a product with the given ID is already in the cart.
•	Added public method removeProduct(int id) that removes a product from the cart entirely.
o	Throws an InvalidParameterException if no product with the given ID is in the cart.
•	Added public method setQuantity(id, newQuantity) that sets the quantity of a given product in the cart to a newly specified value.
o	Throws an InvalidParameterException if no product with the given ID is in the cart.
•	Added public method getQuantity(id) that returns the amount of a given product in the cart. 

StoreView.java
•	Added the StoreView class.
•	Added private attribute store (StoreManager) that contains the storeManager for the store the user is viewing.
•	Added private attribute cartID (int) that contains the cartID that the storeManager assigns the cartID.
•	Added private method displayCart() that displays the current cart and provides the user with all general menu functions as well as those specific to the cart screen.
•	Added private method displayProducts() that displays the store inventory and provides the user with all general menu functions as well as those specific to the inventory screen.
•	Added private method quit() that returns all products in the cart to the store inventory and removes the user’s cart from the list of those in use.
•	Added private method checkout() that displays the total cost of items in the cart and disconnects the user.
•	Added public method viewStore() that provides the user with all general menu functions.
•	UI is fairly limited as it will be replaced with a GUI eventually.
StoreManager.java
•	Added internal class ShoppingCartWithID.
o	This pairs a shopping cart together with its corresponding cartID.
•	Added private attribute shoppingCarts (ArrayList<ShoppingCartWithID>).
o	An ArrayList containing all the carts currently in use and their cartIDs.
o	Carts that have been removed but have cartIDs lower than the highest ID currently in use have their ID set to -1 to mark that they are not in use but still allow the cartID to function as the index of the cart with that ID.
o	Updated StoreManager constructor to initialize this attribute to an empty ArrayList.
•	Added public getInventory() method that returns the inventory.
o	This is intended for use with methods that require the entire inventory. If specific products are needed, use the methods in the next note.
•	Added public getProductName(productID), getProductPrice(productID), and getProductStock(productID) that return the name, price, or stock of a given product.
•	Added public method getCart(cartID) that returns the ShoppingCart associated with the requested product ID.
•	Added public method createCart() that creates a ShoppingCartWithID and adds it to shoppingCarts, then returns the cartID. 
o	This ID will always be the lowest integer currently not in use as a cartID.
•	Added public method removeCart(cartID) that removes the cart with the given ID.
o	If there are carts with higher ID numbers in use, sets the ID to -1 so that the index of those higher ID numbers remains equal to their ID.
	By ensuring that the cartID is equal to the cart’s index in shoppingCarts, accessing specific carts can be O(1) instead of O(N). Creating a cart will take longer, but that only happens once when the cart is created rather than every time an item is added or removed from a cart.
o	If the cart to be removed has the highest ID number in use, removes it and any unused carts that would end up being the new highest ID number.
	This system ensures that the number of unused carts will never be very high by reusing carts before creating new ones and by deleting unused carts wherever possible.
•	Added public method addToCart(cartID, productID, quantity) that attempts to add the requested quantity of the given product to the selected cart. Throws an InvalidParameterException if there is not enough stock to add to cart. If there is, adds it to the cart and removes the amount added to the cart from the inventory.
•	Added public setQuantityInCart(cartID, productID, quantity) that sets a new quantity of a given item in the cart if possible.
o	Returns items to the inventory if the new quantity is lower than the current one.
o	Checks if there is enough in the inventory if the new quantity is higher.
o	Removes the product from the cart if the new quantity is 0.
•	Added public emptyCart(cartID) method that removes all items from a cart and returns them to the inventory, then removes the cart from the list of carts currently in use.
o	Used when a user quits without checking out.
•	Added public processTransaction(cartID) method that removes the user’s cart from the list of carts currently in use and returns the total cost of all items in their cart.
Inventory.java
•	Added public method getInventory() that returns an ArrayList of the product IDs of all the products in the inventory.