Add files via upload
store:
•	Added store package. Inventory, Product, ShoppingCart, Storemanager, and StoreView have been added to this package.
storetest:
•	Added storetest package. Includes TestStoreManager.java, TestInventory.java, and StoreAssertions.java.
testStoreManager.java:
•	Includes unit tests for getInventory, getProductName, getProductPrice, getProductStock, createCart, removeCart, addToCart, removeFromCart, setQuantityInCart, emptyCart, and processTransaction. 
testInventory.java:
•	Includes unit tests for getInventory, getStock, getName, getPrice, addStock, removeStock, and addProduct.
storeAssertions.java:
•	Added StoreAssertions.java.
•	Implements custom assertions for comparing Products and Inventories.
storeexceptions: 
•	Added storeexceptions package. Includes a variety of custom exceptions that are used throughout the project.
•	Includes NegativePriceException, EmptyNameException, NegativeStockException, and InvalidIDException.
ShoppingCart.java
•	Changed all instances of InvalidParameterException to custom exceptions.
o	InvalidIDException thrown in addProduct when attempting to add a product with an ID that is already in use.
o	InvalidIDException thrown in removeProduct when attempting to remove a product that does not exist from the cart.
o	InvalidIDException thrown in setQuantity when attempting to set the quantity of a product that does not exist.
StoreView.java
•	Updated exception handling system.
o	Was previously a number of if statements before most calls to StoreManager methods. These have been replaced with try catch blocks using the new custom exceptions.
StoreManager.java
•	addToCart calls removeStock from Inventory.java, so the exception handling in this method was redundant and has been removed.
•	Changed most instances of InvalidParameterException to custom exceptions.
o	NegativeStockException thrown in setQuantityInCart when attempting to set the quantity of a product in the cart to an amount that would result in the store’s stock of the product going below zero.
•	Removed addStock method. addProductStock exists, making it redundant.
•	getCart method now throws an InvalidIDException instead of an IndexOutOfBoundsException when an ID that is not in use is passed to it.
•	getCart, addToCard, removeFromCart, setQuantityInCart, emptyCart, and processTransaction methods now throw an InvalidIDException if attempting to access a cart that has been removed regardless of whether it is still technically stored. (carts are only actually removed from the ArrayList of carts when they are the final cart in the list and are not in use.)
•	Fixed emptyCart() method not removing carts from the list of active carts.
Inventory.java
•	Changed most instances of InvalidParameterException to custom exceptions.
o	NegativePriceException thrown in addProduct when attempting to add a product with a negative price.
o	EmptyNameException thrown in addProduct when attempting to add a product with an empty string for a name.
o	NegativeStockException thrown in addProduct when attempting to add a product with a negative initial stock.
o	NegativeStockException thrown in changeStock when attempting to reduce a product’s stock below 0.
o	InvalidIDException thrown in findProduct if no product with the given ID exists in the inventory.